Complete test_visualization.py refactoring to meet 600-line limit
Split the 1,111-line test_visualization.py into four focused modules:
- test_visualization_charts.py - Chart generation tests (bar, pie, line, histogram)
- test_visualization_heatmap.py - Heatmap visualization tests
- test_visualization_config.py - Configuration and settings tests
- test_visualization_utils.py - Utility and helper function tests

Updated visualization API to use consistent x_key/y_key parameters and improved
numeric data conversion for better chart generation compatibility.

Fixed CLI tests to use actual file generation instead of mocks for more
realistic integration testing.
